requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
pinecone
tqdm
python-dotenv
python-multipart
requests
numpy